+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Error message in chat when user lose connexion to server
+- Feedback to user when command is executed
 - Command color support
 - Added colors to the usernames
 
@@ -13,6 +15,8 @@
 
 ### Fixed
 
+- 
+- Empty username
 - Fixed hour to print 00:00 format (previous version didn't have leading 0s)
 - Added max limit to the number of messages stored on the server messages array, prevents server memory overload