+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-
+- /roll command
 - Error message in chat when user lose connexion to server
 - Feedback to user when command is executed
 - Command color support